refactor(database)!: change the way the database is sorted
Most importantly now we prioritize end over start.
We also put guild_id as the first parameter for a better.
structure in the sqlite B-tree.

It isn't really a breaking change, since it is still
compatible without errors. However, the sorting will be wrong, and
changing from the previous database to this one requires manually
changing the Primary Key.